# Conflicts:
#	release/datafiles/locale
#	release/datafiles/splash.png
#	release/scripts/addons
#	release/scripts/presets/keyconfig/keymap_data/blender_default.py
#	source/blender/blenkernel/BKE_blender_version.h
#	source/blender/blenkernel/BKE_key.h
#	source/blender/blenkernel/BKE_mesh.h
#	source/blender/blenkernel/intern/key.c
#	source/blender/blenkernel/intern/mesh_normals.cc
#	source/blender/blenkernel/intern/mesh_wrapper.cc
#	source/blender/blenkernel/intern/node_tree_update.cc
#	source/blender/blenkernel/intern/object.cc
#	source/blender/blenloader/intern/versioning_300.c
#	source/blender/bmesh/intern/bmesh_mesh_convert.h
#	source/blender/draw/engines/eevee/eevee_materials.c
#	source/blender/draw/engines/eevee/shaders/closure_eval_lib.glsl
#	source/blender/draw/engines/eevee/shaders/closure_type_lib.glsl
#	source/blender/draw/engines/gpencil/gpencil_cache_utils.c
#	source/blender/draw/engines/overlay/overlay_outline.c
#	source/blender/draw/intern/draw_cache_impl_subdivision.cc
#	source/blender/editors/gizmo_library/gizmo_types/arrow3d_gizmo.c
#	source/blender/editors/include/ED_transform_snap_object_context.h
#	source/blender/editors/io/io_obj.c
#	source/blender/editors/mesh/editmesh_select.c
#	source/blender/editors/transform/transform_gizmo_3d.c
#	source/blender/editors/transform/transform_snap.c
#	source/blender/editors/transform/transform_snap_object.cc
#	source/blender/gpu/GPU_material.h
#	source/blender/gpu/intern/gpu_material.c
#	source/blender/gpu/intern/gpu_material_library.c
#	source/blender/gpu/shaders/gpu_shader_common_obinfos_lib.glsl
#	source/blender/gpu/shaders/material/gpu_shader_material_shader_to_rgba.glsl
#	source/blender/io/wavefront_obj/exporter/obj_export_file_writer.cc
#	source/blender/io/wavefront_obj/exporter/obj_export_file_writer.hh
#	source/blender/io/wavefront_obj/exporter/obj_export_mesh.cc
#	source/blender/io/wavefront_obj/exporter/obj_export_mesh.hh
#	source/blender/io/wavefront_obj/exporter/obj_exporter.cc
#	source/blender/io/wavefront_obj/tests/obj_exporter_tests.cc
#	source/blender/python/bmesh/bmesh_py_types.c
#	source/blender/python/intern/bpy_driver.c
#	source/blender/python/intern/bpy_interface_run.c
#	source/blender/render/intern/multires_bake.c
#	tests/python/bl_pyapi_idprop.py

+++ source/blender/editors/space_text/text_undo.c
@@ -1,18 +1,4 @@
-/*
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- */
+/* SPDX-License-Identifier: GPL-2.0-or-later */
 
 /** \file
  * \ingroup sptext
@@ -73,7 +59,7 @@ typedef struct TextState {
 
 static void text_state_encode(TextState *state, Text *text, BArrayStore *buffer_store)
 {
-  int buf_len = 0;
+  size_t buf_len = 0;
   uchar *buf = (uchar *)txt_to_buf_for_undo(text, &buf_len);
   state->buf_array_state = BLI_array_store_state_add(buffer_store, buf, buf_len, NULL);
   MEM_freeN(buf);
